feat: move utils into separate files
